Fin seccion Security Basico

+++ spring-boot-data-jpa/src/main/java/com/bolsadeideas/springboot/app/controllers/ClienteController.java
@@ -5,6 +5,7 @@ import java.net.MalformedURLException;
 import java.util.Collection;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.apache.juli.logging.Log;
@@ -14,11 +15,14 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestWrapper;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -41,7 +45,7 @@ import com.bolsadeideas.springboot.app.models.service.IUploadFileService;
 @Qualifier("clienteDaoJPA")
 @SessionAttributes("cliente")
 public class ClienteController {
-	
+
 	protected final Log logger = LogFactory.getLog(this.getClass());
 
 	@Autowired
@@ -49,6 +53,7 @@ public class ClienteController {
 	@Autowired
 	private IUploadFileService uploadFileService;
 
+	@Secured({"ROLE_USER","ROLE_ADMIN"})
 	@GetMapping(value = "/uploads/{filename:.+}")
 	public ResponseEntity<Resource> verFoto(@PathVariable String filename) {
 		Resource recurso = null;
@@ -63,6 +68,7 @@ public class ClienteController {
 				.body(recurso);
 	}
 
+	@PreAuthorize("hasRole('ROLE_USER')")
 	@GetMapping(value = "/ver/{id}")
 	public String ver(@PathVariable(value = "id") Long id, Map<String, Object> model, RedirectAttributes flash) {
 		Cliente cliente = clienteService.fetchByIdWithFacturas(id);
@@ -76,30 +82,47 @@ public class ClienteController {
 		return "ver";
 	}
 
-	@GetMapping(value = {"/listar","/"})
-	public String listar(Model model,Authentication authentication) {
-		
-		if(authentication != null) {
+	@GetMapping(value = { "/listar", "/" })
+	public String listar(Model model, Authentication authentication, HttpServletRequest request) {
+
+		if (authentication != null) {
 			logger.info("Hola usuario autenticado, tu username es: ".concat(authentication.getName()));
 		}
-		
+
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		
-		if(auth != null) {
-			logger.info("Utilizando forma estatica SecurityContextHolder.getContext().getAuthentication(): Hola usuario autenticado, tu username es: ".concat(auth.getName()));
+
+		if (auth != null) {
+			logger.info(
+					"Utilizando forma estatica SecurityContextHolder.getContext().getAuthentication(): Hola usuario autenticado, tu username es: "
+							.concat(auth.getName()));
 		}
-		
-		if(hasRole("ROLE_ADMIN")) {
+
+		if (hasRole("ROLE_ADMIN")) {
 			logger.info("Tienes acceso");
-		}else
+		} else
 			logger.info("NO Tienes acceso");
-		
-		
+
+		SecurityContextHolderAwareRequestWrapper securityContext = new SecurityContextHolderAwareRequestWrapper(request,
+				"ROLE_");
+
+		if (securityContext.isUserInRole("ADMIN")) {
+			logger.info("Forma usando SecurityContextHolderAwareRequestWrapper: Tienes acceso");
+		} else {
+			logger.info("Forma usando SecurityContextHolderAwareRequestWrapper: NO Tienes acceso");
+		}
+
+		if (request.isUserInRole("ROLE_ADMIN")) {
+			logger.info("Forma usando HttpServletRequest: Tienes acceso");
+		} else {
+			logger.info("Forma usando HttpServletRequest: NO Tienes acceso");
+		}
+
 		model.addAttribute("titulo", "Listado de clientes");
 		model.addAttribute("clientes", clienteService.findAll());
 		return "listar";
 	}
 
+	@Secured("ROLE_ADMIN")
 	@RequestMapping(value = "/form")
 	public String crear(Map<String, Object> model) {
 		Cliente cliente = new Cliente();
@@ -108,6 +131,7 @@ public class ClienteController {
 		return "form";
 	}
 
+	@Secured("ROLE_ADMIN")
 	@RequestMapping(value = "/form", method = RequestMethod.POST)
 	public String guardar(@Valid @ModelAttribute("cliente") Cliente cliente, BindingResult result, Model model,
 			@RequestParam("file") MultipartFile foto, RedirectAttributes flash, SessionStatus status) {
@@ -141,6 +165,7 @@ public class ClienteController {
 		return "redirect:/listar";
 	}
 
+	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/form/{id}", method = RequestMethod.GET)
 	public String editar(@PathVariable(value = "id") Long id, Map<String, Object> model, RedirectAttributes flash) {
 		Cliente cliente = null;
@@ -162,6 +187,7 @@ public class ClienteController {
 		return "form";
 	}
 
+	@Secured("ROLE_ADMIN")
 	@RequestMapping(value = "/eliminar/{id}", method = RequestMethod.GET)
 	public String eliminar(@PathVariable(value = "id") Long id, RedirectAttributes flash) {
 		if (id > 0) {
@@ -170,38 +196,40 @@ public class ClienteController {
 			clienteService.delete(id);
 			flash.addFlashAttribute("success", "Cliente eliminado con exito");
 
-			if (uploadFileService.delete(cliente.getFoto()))
-				flash.addFlashAttribute("info", "Foto " + cliente.getFoto() + " eliminada con exito!");
+			if (cliente.getFoto() != null)
+				if (uploadFileService.delete(cliente.getFoto()))
+					flash.addFlashAttribute("info", "Foto " + cliente.getFoto() + " eliminada con exito!");
 		}
 
 		return "redirect:/listar";
 	}
-	
+
 	private boolean hasRole(String role) {
 		SecurityContext context = SecurityContextHolder.getContext();
-		
-		if(context == null) {
+
+		if (context == null) {
 			return false;
 		}
-		
+
 		Authentication auth = context.getAuthentication();
-		
-		if(auth == null) {
+
+		if (auth == null) {
 			return false;
 		}
-		
+
 		Collection<? extends GrantedAuthority> authorities = auth.getAuthorities();
-		
+
 		return authorities.contains(new SimpleGrantedAuthority(role));
-		/*for (GrantedAuthority grantedAuthority : authorities) {
-			if(role.equals(grantedAuthority.getAuthority())) {
-				logger.info("Hola usuario " +auth.getName()+ " tu role es: " +grantedAuthority.getAuthority());
-				return true;
-			}
-				
-		}
-		
-		return false;*/
+		/*
+		 * for (GrantedAuthority grantedAuthority : authorities) {
+		 * if(role.equals(grantedAuthority.getAuthority())) {
+		 * logger.info("Hola usuario " +auth.getName()+ " tu role es: "
+		 * +grantedAuthority.getAuthority()); return true; }
+		 * 
+		 * }
+		 * 
+		 * return false;
+		 */
 	}
 
 }
